requirements.txt
# Core
streamlit>=1.40.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pyyaml>=6.0.0

# API
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
python-multipart>=0.0.9

# LLM - LangChain
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.3.0
langchain-google-genai>=2.0.0
langchain-groq>=0.2.0

# Observability
langfuse>=3.0.0

# Document processing
PyMuPDF>=1.24.0
Pillow>=10.0.0
reportlab>=4.0.0
numpy>=1.24.0

# Vector Store
chromadb>=0.5.0

# Visualization
plotly>=5.20.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0
